requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
tenacity>=8.0.0
matplotlib>=3.7.0
yfinance>=0.2.0
ta>=0.11.0
requests>=2.31.0
